criteria WIP : Company nearly done

+++ mavenComputerdatabase/computerDatabaseMaven/src/main/java/com/excilys/mars2020/cdb/service/CompanyService.java
@@ -81,7 +81,7 @@ public class CompanyService {
 	 * 
 	 * @return the number of company in the database
 	 */
-	public int getCountCompanies() {
+	public long getCountCompanies() {
 		return compdao.countAllCompanies();
 	}
